Add support for new Context object
This adds the new context object to the broker api as part of v2.12. This begins to move CF
specific fields from the spec and move them into a context object, for
now they are duplicated. A Cloud Foundry implementation of this will include
org and space guids.

See https://github.com/openservicebrokerapi/servicebroker/issues/115